extract TransparencyJsonWriter
